First attempt at splitting the configuration into two profile so we can configure different beans and config for BrowserStack.
