+++ src/test/java/uk/gov/dhsc/htbhf/utils/NoopWireMockManager.java
@@ -1,0 +1,8 @@
+package uk.gov.dhsc.htbhf.utils;
+
+/**
+ * Implementation of WireMockManager that does nothing which can be used for BrowserStack testing
+ * without having to have conditional logic in the steps.
+ */
+public class NoopWireMockManager implements WireMockManager {
+}
