+++ src/test/resources/application-browserstack.properties
@@ -1,0 +1,2 @@
+base.url=https://help-to-buy-healthy-foods-staging.london.cloudapps.digital
+session.details.url=https://help-to-buy-healthy-foods-staging.london.cloudapps.digital/session-details/confirmation-code
