+++ src/test/resources/application-local.properties
@@ -1,0 +1,5 @@
+test.browser=chrome
+base.url=http://localhost:8080
+test.headless=true
+session.details.url=http://localhost:8081/session-details/confirmation-code
+wiremock.port=8090
