+++ src/test/resources/application.properties
@@ -1,6 +1,2 @@
-test.browser=chrome
-base.url=http://localhost:8080
 wait.timeout.seconds=15
-test.headless=true
-session.details.url=http://localhost:8081/session-details/confirmation-code
-wiremock.port=8090
+spring.profiles.active=local
